backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pydantic==2.6.0
pyjwt==2.8.0
bcrypt==4.1.2
requests==2.31.0
python-multipart==0.0.9
email-validator==2.1.0.post1
